fix: Moving misplaced function
